+++ tw-tasks-core/src/main/java/com/transferwise/tasks/handler/SimpleTaskProcessingPolicy.java
@@ -6,6 +6,7 @@ import com.transferwise.tasks.buckets.IBucketsManager;
 import com.transferwise.tasks.domain.IBaseTask;
 import com.transferwise.tasks.handler.interfaces.ITaskProcessingPolicy;
 import com.transferwise.tasks.handler.interfaces.StuckDetectionSource;
+import com.transferwise.tasks.helpers.kafka.partitionkey.IPartitionKeyStrategy;
 import java.time.Duration;
 import java.time.Instant;
 import lombok.Getter;
@@ -36,6 +37,11 @@ public class SimpleTaskProcessingPolicy implements ITaskProcessingPolicy {
   @Setter
   @Accessors(chain = true)
   private StuckTaskResolutionStrategy stuckTaskResolutionStrategy = StuckTaskResolutionStrategy.MARK_AS_ERROR;
+
+  @Getter
+  @Setter
+  @Accessors(chain = true)
+  private IPartitionKeyStrategy partitionKeyStrategy = null;
 
   @Override
   public String getProcessingBucket(IBaseTask task) {